refactor(broker): correct import path to new broker-client
